Reorganises project structure so that model is in its own module and all services are in the same module.
