+++ model/src/main/java/com/google/gfx/LocationMailingList.java
@@ -1,0 +1,15 @@
+package com.google.gfx;
+
+/**
+ * An email address that is notified for every new order.
+ */
+public class LocationMailingList {
+
+  private String emailAddress;
+  private Location location;
+
+  public LocationMailingList(String emailAddress, Location location) {
+    this.emailAddress = emailAddress;
+    this.location = location;
+  }
+}
