+++ services/discovery/pom.xml
