--- src/main/java/com/google/gfx/model/LocationMailingList.java
@@ -1,15 +1,0 @@
-package com.google.gfx.model;
-
-/**
- * An email address that is notified for every new order.
- */
-public class LocationMailingList {
-
-  private String emailAddress;
-  private Location location;
-
-  public LocationMailingList(String emailAddress, Location location) {
-    this.emailAddress = emailAddress;
-    this.location = location;
-  }
-}
